Settings: SuperSU from Settings
If SuperSU is installed, show an entry in Settings below
"Developer options", which launches SuperSU when tapped.
At boot time SuperSU will detect if com.android.settings
contains the resources from this patch, and if so it
will hide itself from the launcher. This was implemented
in v1.85, so we only show the entry if the installed
SuperSU version is at least that.

PS2: Better Drawables courtesy of Lo

Signed-off-by: ayysir <dresadd09691@gmail.com>
Signed-off-by: houst0nn <houstonn@gmail.com>
Change-Id: I1796d92458587b9e961d5d8e6d58099e8367834d

Conflicts:
	res/xml/settings_headers.xml