Settings: Wise Order Settings
Android during the years has improved it's design and guidelines. This change is meant to be a start point to redefine how settings are handled.

Current Status:

- Give case sensitive to Section notes to improve their reading
- Move location settings to Networking
- Personal settings before Device settings
- Add advanced section (Users need to realize that some information could be misunderstood under that section)

WIP:

The overall Android system needs to be polished in future releases. Common users don't waste their time browsing each section or fragment. The right information needs to be shown on top of less important settings. That means less settings but nearer to the end-user.

Change-Id: Idbd878ccfda4e21e062be504e4529681a3949d8c
Conflicts:
	res/xml/settings_headers.xml